requirements.txt
Babel==2.9.1
mpmath==1.2.1
Pillow==8.4.0
prettytable==2.4.0
psycopg2-binary==2.9.2
pygame==2.1.0
pytz==2021.3
sympy==1.9
tkcalendar==1.6.1
wcwidth==0.2.5
